# specs/005-user-org-service/contracts/user-org-service.openapi.yaml
openapi: 3.1.0
info:
  title: User & Organization Service API
  version: 0.1.0
  description: >
    Contracts for interactive admin API, authentication flows, policy decisions,
    budget enforcement, declarative reconciliation controls, and audit exports.
servers:
  - url: https://user-org-service.{env}.platform.internal
    description: Cluster-scoped service endpoint (dev/staging/prod)
  - url: http://localhost:8081
    description: Local development (admin API)
...
security:
  - bearerAuth: []
  - m2mToken: []
paths:
  /v1/auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate an interactive user with password and MFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '423':
          description: Account locked
  /v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate access/refresh tokens
      security:
        - refreshToken: []
      responses:
        '200':
          description: Tokens rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Refresh token with dedicated scope
    m2mToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Machine-to-machine token issued to downstream services
...
  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        totp:
          type: string
          description: Required when MFA enforced
        webauthnAssertion:
          type: object
          description: Used for WebAuthn completion
    AuthSession:
      type: object
      required: [accessToken, refreshToken, sessionId, expiresAt]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        sessionId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
        mfaVerified:
          type: boolean
...